+++ example/lib/pages/normal_player_page.dart
@@ -19,20 +19,12 @@ class _NormalPlayerPageState extends State<NormalPlayerPage> {
       fit: BoxFit.contain,
     );
     BetterPlayerDataSource dataSource = BetterPlayerDataSource(
-        BetterPlayerDataSourceType.network, Constants.forBiggerBlazesUrl);
+        BetterPlayerDataSourceType.network,
+        "https://v5.szjal.cn/20200829/qByymCNH/index.m3u8");
     _betterPlayerController = BetterPlayerController(betterPlayerConfiguration);
     _betterPlayerController.setupDataSource(dataSource);
 
     super.initState();
-    // _betterPlayerController.setControlsEnabled(false);
-    _betterPlayerController.addEventsListener((event) {
-      if (event.betterPlayerEventType == BetterPlayerEventType.initialized) {
-        _betterPlayerController.setControlsEnabled(false);
-      }
-      if (event.betterPlayerEventType == BetterPlayerEventType.openFullscreen) {
-        _betterPlayerController.setControlsEnabled(true);
-      }
-    });
   }
 
   @override