+++ lib/better_player.dart
@@ -21,3 +21,4 @@ export 'src/playlist/better_player_playlist_configuration.dart';
 export 'src/subtitles/better_player_subtitles_configuration.dart';
 export 'src/subtitles/better_player_subtitles_source.dart';
 export 'src/subtitles/better_player_subtitles_source_type.dart';
+export 'src/video_player/video_player.dart' show VideoPlayerValue;